requirements.txt
# LLM & RAG
openai>=1.30
langchain>=0.2
langchain-openai>=0.1
langchain-community>=0.2

# Vector store
chromadb>=0.5

# Sparse retrieval
rank-bm25>=0.2

# Evaluation
ragas>=0.1
datasets>=2.19       # required by ragas

# Re-ranking (optional but recommended)
sentence-transformers>=3.0

# UI
streamlit>=1.35

# Data
pandas>=2.2
pyarrow>=15.0
requests>=2.31

# Utils
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.66

# Dev / testing
pytest>=8.0
pytest-cov>=5.0
ruff>=0.4
beautifulsoup4>=4.12